+++ src/util/const.ts
@@ -55,11 +55,38 @@ export const EthersNetwork = {
   [Network.ASTAR_MAINNET]: 'astar-mainnet',
   [Network.POLYGONZKEVM_MAINNET]: 'polygonzkevm-mainnet',
   [Network.POLYGONZKEVM_TESTNET]: 'polygonzkevm-testnet',
+  [Network.POLYGONZKEVM_CARDONA]: 'polygonzkevm-cardona',
   [Network.BASE_MAINNET]: 'base-mainnet',
   [Network.BASE_GOERLI]: 'base-goerli',
   [Network.BASE_SEPOLIA]: 'base-sepolia',
   [Network.ZKSYNC_MAINNET]: 'zksync-mainnet',
-  [Network.ZKSYNC_SEPOLIA]: 'zksync-sepolia'
+  [Network.ZKSYNC_SEPOLIA]: 'zksync-sepolia',
+  [Network.SHAPE_MAINNET]: 'shape-mainnet',
+  [Network.SHAPE_SEPOLIA]: 'shape-sepolia',
+  [Network.LINEA_MAINNET]: 'linea-mainnet',
+  [Network.LINEA_SEPOLIA]: 'linea-sepolia',
+  [Network.FANTOM_MAINNET]: 'fantom-mainnet',
+  [Network.FANTOM_TESTNET]: 'fantom-testnet',
+  [Network.ZETACHAIN_MAINNET]: 'zetachain-mainnet',
+  [Network.ZETACHAIN_TESTNET]: 'zetachain-testnet',
+  [Network.ARBNOVA_MAINNET]: 'arbnova-mainnet',
+  [Network.BLAST_MAINNET]: 'blast-mainnet',
+  [Network.BLAST_SEPOLIA]: 'blast-sepolia',
+  [Network.MANTLE_MAINNET]: 'mantle-mainnet',
+  [Network.MANTLE_SEPOLIA]: 'mantle-sepolia',
+  [Network.SCROLL_MAINNET]: 'scroll-mainnet',
+  [Network.SCROLL_SEPOLIA]: 'scroll-sepolia',
+  [Network.GNOSIS_MAINNET]: 'gnosis-mainnet',
+  [Network.GNOSIS_CHIADO]: 'gnosis-chiado',
+  [Network.BNB_MAINNET]: 'bnb-mainnet',
+  [Network.BNB_TESTNET]: 'bnb-testnet',
+  [Network.AVAX_MAINNET]: 'avax-mainnet',
+  [Network.AVAX_FUJI]: 'avax-fuji',
+  [Network.CELO_MAINNET]: 'celo-mainnet',
+  [Network.CELO_ALFAJORES]: 'celo-alfajores',
+  [Network.METIS_MAINNET]: 'metis-mainnet',
+  [Network.OPBNB_MAINNET]: 'opbnb-mainnet',
+  [Network.OPBNB_TESTNET]: 'opbnb-testnet'
 };
 
 /**
@@ -96,6 +123,10 @@ export const CustomNetworks: { [key: string]: NetworkFromEthers } = {
     chainId: 1442,
     name: 'polygonzkevm-testnet'
   },
+  'polygonzkevm-cardona': {
+    chainId: 2442,
+    name: 'polygonzkevm-cardona'
+  },
   'base-mainnet': {
     chainId: 8453,
     name: 'base-mainnet'
@@ -119,6 +150,110 @@ export const CustomNetworks: { [key: string]: NetworkFromEthers } = {
   'zksync-sepolia': {
     chainId: 300,
     name: 'zksync-sepolia'
+  },
+  'shape-mainnet': {
+    chainId: 360,
+    name: 'shape-mainnet'
+  },
+  'shape-sepolia': {
+    chainId: 11011,
+    name: 'shape-sepolia'
+  },
+  'linea-mainnet': {
+    chainId: 59144,
+    name: 'linea-mainnet'
+  },
+  'linea-sepolia': {
+    chainId: 59141,
+    name: 'linea-sepolia'
+  },
+  'fantom-mainnet': {
+    chainId: 250,
+    name: 'fantom-mainnet'
+  },
+  'fantom-testnet': {
+    chainId: 4002,
+    name: 'fantom-testnet'
+  },
+  'zetachain-mainnet': {
+    chainId: 7000,
+    name: 'zetachain-mainnet'
+  },
+  'zetachain-testnet': {
+    chainId: 7001,
+    name: 'zetachain-testnet'
+  },
+  'arbnova-mainnet': {
+    chainId: 42170,
+    name: 'arbnova-mainnet'
+  },
+  'blast-mainnet': {
+    chainId: 81457,
+    name: 'blast-mainnet'
+  },
+  'blast-sepolia': {
+    chainId: 168587773,
+    name: 'blast-sepolia'
+  },
+  'mantle-mainnet': {
+    chainId: 5000,
+    name: 'mantle-mainnet'
+  },
+  'mantle-sepolia': {
+    chainId: 5003,
+    name: 'mantle-sepolia'
+  },
+  'scroll-mainnet': {
+    chainId: 534352,
+    name: 'scroll-mainnet'
+  },
+  'scroll-sepolia': {
+    chainId: 534351,
+    name: 'scroll-sepolia'
+  },
+  'gnosis-mainnet': {
+    chainId: 100,
+    name: 'gnosis-mainnet'
+  },
+  'gnosis-chiado': {
+    chainId: 10200,
+    name: 'gnosis-chiado'
+  },
+  'bnb-mainnet': {
+    chainId: 56,
+    name: 'bnb-mainnet'
+  },
+  'bnb-testnet': {
+    chainId: 97,
+    name: 'bnb-testnet'
+  },
+  'avax-mainnet': {
+    chainId: 43114,
+    name: 'avax-mainnet'
+  },
+  'avax-fuji': {
+    chainId: 43113,
+    name: 'avax-fuji'
+  },
+  'celo-mainnet': {
+    chainId: 42220,
+    name: 'celo-mainnet'
+  },
+  'celo-alfajores': {
+    chainId: 44787,
+    name: 'celo-alfajores'
+  },
+  'metis-mainnet': {
+    chainId: 1088,
+    name: 'metis-mainnet'
+  },
+  'opbnb-mainnet': {
+    chainId: 204,
+    name: 'opbnb-mainnet'
+  },
+  'opbnb-testnet': {
+    chainId: 5611,
+    name: 'opbnb-testnet'
   }
 };